requirements.txt
agno
browser-use
google-genai
numpy
streamlit
langchain-google-genai
langchain-openai
langchain-anthropic
langchain-groq
playwright
beautifulsoup4
aiohttp
python-dotenv